refactor : 인터페이스 ->타입으로 변경
